🔥 refactor(Coach): Remove salary and address fields from CoachVO 🔥 refactor(Goods): Remove unitPrice field from GoodsVO 🔥 refactor(GoodsTransactions): Remove memberPhone field from GoodsTransactionsVO 🗑️ chore(Entities): Delete unused entity and mapper files ✨ feat(Equipment): Add status field to EquipmentQueryRequest ✨ feat(Course): Add category, difficulty and image fields ✨ feat(CoursePurchase): Add status filter to query ✨ feat(GoodsTransactions): Add memberName filter to query ✨ feat(Goods): Add isDelete filter to query 🔧 fix(Controllers): Update controller methods and annotations 🔧 fix(Services): Implement enhanced course service logic

+++ backend/src/main/java/com/liucf/gymsystembackend/controller/GoodsController.java
@@ -38,7 +38,7 @@ public class GoodsController {
      */
     @PostMapping("/add")
     @AuthCheck(mustRole = MemberConstant.ADMIN_ROLE)
-    public BaseResponse<Long> addGoods(@RequestBody GoodsAddRequest goodsAddRequest) {
+    public BaseResponse<Integer> addGoods(@RequestBody GoodsAddRequest goodsAddRequest) {
         ThrowUtils.throwIf(goodsAddRequest == null, ErrorCode.PARAMS_ERROR);
         Goods goods = new Goods();
         BeanUtils.copyProperties(goodsAddRequest, goods);
